Vegetables Remaining Box
BUILDS ON ABEL'S BACKEND COMMIT

Added a box to record the vegetables remaining on the board. Added a global variable to store the carrots remaining (defaulting at 2 for the 3x3). Added a function to update the number of carrots remaining - called on start (to fix default value) and to be called when carrot revealed.