Question Pop-Up UI implementation
Added the UI elements with images, text and input text.

Attempted to add the gameobject to the 3x3 script but it returns an object error which I believe is something to do with the QuestionPopUp not being initialized in the script but im unsure how to define it globally.

Added to the Question function to make the QuestionPopUp object Active when in the "InQuestion" game mode.